Fix merge issue mixing Client and ShelterVolunteer
