+++ src/main/java/cat/udl/eps/softarch/demo/domain/ShelterVolunteer.java
@@ -1,21 +1,5 @@
 package cat.udl.eps.softarch.demo.domain;
 
-import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Entity;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
+public class ShelterVolunteer extends User {
 
-import java.util.Collection;
-
-@EqualsAndHashCode(callSuper = true)
-@Entity
-@Data
-public class Client extends User {
-    @Override
-    @ElementCollection
-    public Collection<GrantedAuthority> getAuthorities(){
-        return AuthorityUtils.commaSeparatedStringToAuthorityList("ROLE_CLIENT");
-    }
 }
